Change Http.Service.request from UIO to Task (#147)
The `Http.live` implementation uses a lot of the `java.net._` API, which
throws all kinds of exceptions.  This changes the effect type
accordingly, so that if something goes wrong we can decide how to handle
it from within the `Producer` implementation.  For now, we shrug it off
and do nothing.