Add support for partial environment injection
This allows a `Has[R0]` to be injected into a `ZIO[Has[R0] with Has[R1],
E, A]`, yielding a `Zio[Has[R1], E, A]`.

This change allows `Clock` instances to be excluded from the explicit
environment provided in test code, since zio-test already includes that
in the test environment.