Add support for SQLite persistence, @count (#157)
This adds a new database service `Db` that wraps SQLite, and a producer
that responds to `@count` as an example of how to use it.

It's currently up to each producer to deal with isolation from other
producers, and between channels and nicks, as well as transaction
management.  These are all potential areas for improvement of the
`Db` service.