Add support for looking up the weather report
This adds support for `@wx <location>`, which gets the current weather
at the given location.

OpenStreetMap is used to search for the location, and DarkSky is used to
get weather by lat/lon.  Json4s is used for parsing JSON responses from
both.  DarkSky requires an API key, which must be set in the
`DARK_SKY_API_KEY` environment variable.