Move CoinDesk sample response to a .json file (#188)
This tidies up the @btc test code by moving the big block of JSON from
the Scala source to a *.json* file under src/test/resources.  The
*.json* file includes the full response from the CoinDesk API, rather
than the truncated version in *BtcSpec.scala*, which makes for a more
accurate test and will be useful for future reference.

This also adds a utility to read in a string given a path on the
classpath.